pyproject.toml
[project]
name = "excellingo"
version = "0.1.0"
description = "ExcelLingo - 智能表格翻译工具"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.0",
    "tkinter-tooltip>=2.0.0",
    "aiohttp>=3.9.0",
    "asyncio-throttle>=1.0.0",
    "tkinterdnd2>=0.3.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]